code refactor, apply  separate of concern
