+++ src/components/hoc/contextProvider/index.ts
@@ -1,1 +1,2 @@
+export { AppProvider } from './AppProvider';
 export { ChatProvider } from './ChatProvider';
